refactor: Move gamecluster to pkg level
