+++ cmd/avalond/avalond.go
@@ -13,11 +13,11 @@ import (
 	"github.com/0xa1-red/empires-of-avalon/actor/timer"
 	"github.com/0xa1-red/empires-of-avalon/config"
 	"github.com/0xa1-red/empires-of-avalon/database"
-	"github.com/0xa1-red/empires-of-avalon/gamecluster"
 	"github.com/0xa1-red/empires-of-avalon/instrumentation/metrics"
 	"github.com/0xa1-red/empires-of-avalon/instrumentation/traces"
 	"github.com/0xa1-red/empires-of-avalon/logging"
 	"github.com/0xa1-red/empires-of-avalon/persistence"
+	gamecluster "github.com/0xa1-red/empires-of-avalon/pkg/cluster"
 	"github.com/0xa1-red/empires-of-avalon/pkg/service/auth"
 	"github.com/0xa1-red/empires-of-avalon/pkg/service/blueprints"
 	"github.com/0xa1-red/empires-of-avalon/pkg/service/registry"
